% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpj_stats.R
\name{lpj_stats}
\alias{lpj_stats}
\title{Run summary statistics (mean, min, range, sd, etc.) on all variables in an output file.
Choose whether to aggregate by year, gridcell, or both.}
\usage{
lpj_stats(nc_path, lvl = "both")
}
\arguments{
\item{nc_path}{string. file path to LPJG netCDF output}

\item{lvl}{string. one of: "year", "gridcell", "both". defaults to "both".}
}
\value{
a tibble
}
\description{
Run summary statistics (mean, min, range, sd, etc.) on all variables in an output file.
Choose whether to aggregate by year, gridcell, or both.
}
