% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lpj.R
\name{read_lpj}
\alias{read_lpj}
\title{read_lpj}
\usage{
read_lpj(nc_path, var_name, res, start = "1901-01-01")
}
\arguments{
\item{nc_path}{path to netcdf file containing lpj output}

\item{var_name}{variable(s) to extract from nc file ("var1" or c("var1", "var2", ...))}

\item{res}{temporal resolution of nc file ("year", "month", or "day")}

\item{start}{gets passed to seq_date_lpj to determine beginning of time series}
}
\value{
returns a nested tibble grouped by pfts or pfts and vars for one or more
    var inputs, respectively
}
\description{
read_lpj
}
\examples{
\dontrun{
path <- "path/to/lpj/nc/file.nc"
tidy_output <- read_lpj(path, c("Pft-Out/npp", "Pft-Out/gpp"), "year")
}

}
