% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lpj_meta.R
\name{read_lpj_meta}
\alias{read_lpj_meta}
\title{read_lpj_meta returns the names of all variables contained in an LPJ-GUESS .nc output file}
\usage{
read_lpj_meta(nc_path)
}
\arguments{
\item{nc_path}{}
}
\value{
a tibble
}
\description{
read_lpj_meta returns the names of all variables contained in an LPJ-GUESS .nc output file
}
