% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lpj_patch.R
\name{read_lpj_patch}
\alias{read_lpj_patch}
\title{read_lpj_patch}
\usage{
read_lpj_patch(nc_path, var_name)
}
\arguments{
\item{nc_path}{path to netcdf file containing lpj output}

\item{var_name}{variable(s) to extract from nc file ("var1" or c("var1", "var2", ...))}
}
\value{
a tibble
}
\description{
read_lpj_patch
}
\examples{
\dontrun{
path <- "path/to/lpj/nc/file.nc"
tidy_output <- read_lpj_pft(path, c("Pft-Out/npp", "Pft-Out/gpp"))
}

}
